Module imbriqua_structure_loader::module_log

Configure resilient logger : this module provide logger config using two configuration

How to use

Configure logger with YAML configuration file, after configuring logger with backup configuration (described in next section)

This two-step provide a working detailed logging backup configuration during configuration file error

Minimal usecase

mod module_log;
use log::{error, warn, info, debug, trace};
 
fn main() {
 
    let _handle = module_log::open_module();
 
    error!("It's an error log");
    warn!("It's a warn log");
    info!("It's an info log");
    debug!("It's a debug log");
    trace!("It's a trace log");
 
}

Customization

Edit config_log.yaml file configuration this log4rs notation

Example config_log.yaml file configuration

appenders:
    stdout:
        kind: console
        encoder:
            pattern: "{M}: {m} {n}"
        filters:
            - kind: threshold
              level: info
 
    requests:
        kind: file
        path: "imbriqua_structure_loader.log"
        encoder:
            pattern: "{d(%+)(utc)} [{f}:{L}] {h({l})} {M}: {m} {n}"
        filters:
            - kind: threshold
              level: trace
 
root:
    level: trace
    appenders:
        - stdout
        - requests

Equivalent YAML file of backup configuration

appenders:
    stdout:
        kind: console
        encoder:
            pattern: "! BACKUP_LOGGER ! {M}: {m} {n}"
        filters:
            - kind: threshold
              level: trace
 
    requests:
        kind: file
        path: "imbriqua_structure_loader.log"
        encoder:
            pattern: "{d(%+)(utc)} [{f}:{L}] {h({l})} ! BACKUP_LOGGER ! {M}: {m} {n}"
        filters:
            - kind: threshold
              level: trace
 
root:
    level: trace
    appenders:
        - stdout
        - requests

Panic and failure

WARN_LOG01 - Error during default configuration loading

Context : module_log.rs/open_modules()

Info : The file config_log.yml can’t be loaded in log4rs configuration

Info : A backup logging configuration may be load,logs are in imbriqua_structure_loader.log file

Cause : See logs for syntaxe error details, or deserialize error details

PANIC_LOG01 - Error during the loading on logs modules

Context : module_log.rs/open_modules()

Info : No logs are provided, so, make panic

Cause : The error come from get_config_by_backup() function

PANIC_LOG02 - Error during the loading on logs modules

Context : module_log.rs/open_modules()

Info : No logs are provided, so, make panic

Cause : The error come from log4rs::init_config() function

Functions

get_config_by_backup 🔒

Define a backup config (hard-writted)

get_config_by_file 🔒

Define a config by loading “config_log.yml”

open_module

Configure logger with config_log.yml, after configure it with a backup configuation